backend/requirements.txt
coverage==7.3.1
Django==4.2.5
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.7
factory-boy==3.3.0
gunicorn==21.2.0
isort==5.12.0
oauthlib==3.2.2
pandas==2.1.1
psycopg2-binary==2.9.8
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
